Read file with Classic asp
Supporting legacy code, which is why there will be a lot to upload.

This is currently a poor file structure, but it will change as I perform
my tests.

Normally you would not keep commits close together like this, but I have
reasons. (Time is a big one). A good portion of this test is pulling to
another operating system and back again.